refactor: enhancing program
