+++ programs/fee_collector/src/errors.rs
